feat(collapse): Add collapse component
The commit adds a new collapse component `@fx-lib/collapse`
package. The collapse component is a versatile and interactive component
that allows creating collapsible sections in web applications,
conserving space, and enhancing the user experience. The component
supports vertical and horizontal orientations and provides options for
the default state, such as open or close.