Refactor modal helper functions and introduce modal overlay feature
This commit refactors the modal helper functions in order to improve
code organization and readability. The `appendBefore` function is
removed and replaced with the `buildOverlay` and `destroyOverlay`
functions, which handle the creation and removal of the modal overlay
respectively. The modal overlay is now a separate file
`modalOverlay.ts`. This change allows for greater control over the modal
overlay element and enhances the reusability