Refactor tab trigger and tab panel attributes to use data-state instead of aria-selected to avoid accessibility issue
This change updates the tab trigger attributes from "aria-selected" to
"data-state" in order to track the state of the triggers accurately. The
"data-state" attribute now stores the values "active" and "inactive" to
indicate the active and inactive states, respectively. This modification
allows for more efficient and precise management of the tab triggers.
Aria-selected has been replaced for accessibility since button or tabs panel are not supporting aria-selected attribute.

Issue: #11