feat(dropdown): Improve dropdown item focus handling
This change enhances the handling of dropdown item focus in the Dropdown
class. The `hide()` method of the Dropdown class now removes focus from
the active item when the dropdown is hidden. This ensures that the focus
is properly managed and prevents issues with retained focus when the
dropdown is reopened.

Additionally, refactor `injectStyles` mode in the Dropdown, This eliminates many call, now this is an independant feature, can be called if needed of create custom styles.

These changes improve the overall usability and stability of the
dropdown component.